Had to modify various classes in the food use case to use float instead of double values

+++ src/use_case/food/FoodDataAccessInterface.java
@@ -4,7 +4,7 @@ import java.util.HashMap;
 
 public interface FoodDataAccessInterface {
 
-    HashMap<String, Double> getFoodNutritionalValues(String foodName, Float quantity);
+    HashMap<String, Float> getFoodNutritionalValues(String foodName, Float quantity);
 
     Integer getFoodCalories(String foodName, Float quantity);
 //    To-do: Decide whether the calory related data is necessary
